Proper MVC Architecture file and folder structure:  controller, model, service, and view separation is well-implemented
Maven Project Structure: Standard Maven directory layout followed correctly

Package Organization for Good separation of concerns with distinct packages

motorph_payroll_system/
├── data/                          # 📁 Data files (moved from root)
│   ├── attendanceRecord.csv
│   └── employeeDetails.csv
├── pom.xml                        # Maven configuration
└── src/
    ├── main/java/com/motorph/     # 📁 Main application code
    │   ├── Main.java              # Application entry point
    │   ├── controller/            # 🎮 MVC Controllers
    │   │   ├── EmployeeController.java
    │   │   ├── PayrollController.java
    │   │   └── ReportController.java
    │   ├── model/                 # 📋 Data Models
    │   │   ├── AttendanceRecord.java
    │   │   ├── Employee.java
    │   │   └── PaySlip.java
    │   ├── repository/            # 💾 Data Access (cleaned)
    │   │   └── DataRepository.java
    │   ├── service/               # ⚙️ Business Logic
    │   │   ├── EmployeeService.java
    │   │   ├── PayrollProcessor.java
    │   │   ├── PayrollService.java
    │   │   └── ReportService.java
    │   ├── util/                  # 🔧 Utilities
    │   │   ├── DateUtils.java
    │   │   ├── ErrorHandler.java
    │   │   ├── InputValidator.java
    │   │   ├── PayrollConstants.java
    │   │   ├── UIConstants.java
    │   │   └── UIUtils.java
    │   └── view/                  # 🖼️ User Interface
    │       ├── ApplicationMenuBar.java
    │       ├── EmployeeListPanel.java    # (unified)
    │       ├── EmployeeManagementPanel.java
    │       ├── MainFrame.java
    │       ├── MainMenuPanel.java
    │       ├── PayrollPanel.java
    │       ├── ReportsPanel.java
    │       ├── dialog/            # Dialog windows
    │       │   ├── DateRangeDialog.java
    │       │   ├── EditEmployeeDialog.java
    │       │   ├── EmployeeDetailsFrame.java
    │       │   ├── EmployeeNumberInputDialog.java
    │       │   ├── NewEmployeeDialog.java
    │       │   ├── PayslipDialog.java
    │       │   └── SearchResultDialog.java
    │       └── renderer/          # Custom UI renderers
    │           └── ActionButtonRenderer.java
    └── test/java/com/motorph/     # 🧪 Test files (properly organized)
        ├── CSVCreateAndWrite.java
        ├── CSVTest.java
        ├── CSVWriteTest.java
        ├── CredentialManager.java
        ├── EmployeeServiceTest.java
        ├── ManualCSVTest.java
        └── SimpleCSVTest.java
