+++ motorph_payroll_system/data/employeeDetails.csv
@@ -33,4 +33,3 @@
 "10032","Castro","John Rafael","02/09/1992","Hi-way, Yati, Liloan Cebu","332-424-955 ","26-7145133-4","601644902402","404-768-309-000","697764069311","Regular","Sales & Marketing","Reyes, Isabella","""52,670""","""1,500""","""1,000""","""1,000""","""26,335""","313.51190476190476"
 "10033","Martinez","Carlos Ian","11/16/1990","Bulala, Camalaniugan","078-854-208","11-5062972-7","380685387212","256-436-296-000","993372963726","Regular","Supply Chain and Logistics","Reyes, Isabella","""52,670""","""1,500""","""1,000""","""1,000""","""26,335""","313.51190476190476"
 "10034","Santos","Beatriz","08/07/1990","Agapita Building, Metro Manila","526-639-511","20-2987501-5","918460050077","911-529-713-000","874042259378","Regular","Customer Service and Relations","Reyes, Isabella","""52,670""","""1,500""","""1,000""","""1,000""","""26,335""","313.51190476190476"
-"999999","GROUP4","CP2","01/01/2001","MMDC","999","999","999","999","999","Regular","student","mentor","""999""","""1,500""","""500""","""500""","""499""","5.946428571428571"
